requirements.txt
django
djangorestframework
django-rest-framework
Jinja2
requests
djutils
pexpect